Byte 8 bit

+++ src/main/java/gregtech/api/interfaces/IToolStats.java
@@ -197,5 +197,5 @@ public interface IToolStats {
 
     String getToolTypeName();
 
-    int getMaxMode();
+    byte getMaxMode();
 }
